requirements.txt
pygame-ce==2.5.7
numpy==2.4.4
